Organize deployment files in subfolders
