--- deploy/artifacts.js
@@ -1,9 +1,0 @@
-var _artifacts;
-
-module.exports = function(contractName) {
-  return _artifacts.require(contractName);
-};
-
-module.exports.setArtifacts = function(val) {
-  _artifacts = val;
-}
