+++ deploy/objects/RegistryManager.js
