+++ deploy/scripts/initProject.js
@@ -1,9 +1,9 @@
 // Create a project controller for your project on top of zOS
 // Run as: `npx truffle exec scripts/init_project.js --network NETWORK`
 
-import Deployer from "../deploy/Deployer";
-import ZepCoreManager from "../deploy/ZepCoreManager";
-import { PROJECT_ACCOUNT, PROJECT_NAME, ZEPPELIN_ACCOUNT, DEVELOPER_ACCOUNT, DEVELOPER_FRACTION, NEW_VERSION_COST, VERSION, DISTRIBUTION, ERC721_CONTRACT_NAME } from '../deploy/constants'
+import Deployer from "../objects/Deployer";
+import ZepCoreManager from "../objects/ZepCoreManager";
+import { PROJECT_ACCOUNT, PROJECT_NAME, ZEPPELIN_ACCOUNT, DEVELOPER_ACCOUNT, DEVELOPER_FRACTION, NEW_VERSION_COST, VERSION, DISTRIBUTION, ERC721_CONTRACT_NAME } from '../constants'
 
 const ERC721Token = artifacts.require('ERC721Token');
